requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==42.0.5
defusedxml==0.8.0rc2
Django==5.0.6
django-allauth==0.63.3
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-crontab==0.7.1
django-timezone-field==6.1.0
djangorestframework==3.15.1
idna==3.7
jwt==1.3.1
kombu==5.3.7
mysql-connector-python==8.3.0
mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
pandas==2.2.2
pillow==10.3.0
prompt_toolkit==3.0.45
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
PyMySQL==1.1.1
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==70.0.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
